+++ mod-record-specifications-server/src/main/java/org/folio/rspec/controller/handler/MethodArgumentNotValidExceptionHandler.java
@@ -3,8 +3,10 @@ package org.folio.rspec.controller.handler;
 import static org.folio.rspec.controller.handler.ServiceExceptionHandler.fromErrorCode;
 import static org.folio.rspec.domain.dto.ErrorCode.INVALID_REQUEST_PARAMETER;
 
+import com.google.common.collect.Lists;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.bouncycastle.util.Arrays;
 import org.folio.rspec.domain.dto.Error;
 import org.folio.rspec.domain.dto.ErrorCollection;
 import org.folio.rspec.domain.dto.Parameter;
@@ -20,6 +22,8 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 public class MethodArgumentNotValidExceptionHandler implements ServiceExceptionHandler {
 
   private static final String FIELD_MSG_ARG = "field";
+  private static final String BOUNDARY_MSG_ARG = "boundary";
+  private static final String UNDEFINED_VALUE = "undefined";
 
   private final TranslationService translationService;
 
@@ -40,9 +44,27 @@ public class MethodArgumentNotValidExceptionHandler implements ServiceExceptionH
 
   private Error createErrorFromFieldError(FieldError fieldError) {
     var error = fromErrorCode(INVALID_REQUEST_PARAMETER);
-    error.setMessage(translationService.format(fieldError.getCodes(), FIELD_MSG_ARG, fieldError.getField()));
+    error.setMessage(translationService.format(fieldError.getCodes(),
+      FIELD_MSG_ARG, fieldError.getField(),
+      BOUNDARY_MSG_ARG, getBoundary(fieldError)));
     var parameter = new Parameter().key(fieldError.getField()).value(String.valueOf(fieldError.getRejectedValue()));
     error.addParametersItem(parameter);
     return error;
+  }
+
+  /**
+   * Get boundary for Min/Max or minLength/maxLength violations.
+   * */
+  private Object getBoundary(FieldError fieldError) {
+    var args = fieldError.getArguments();
+    if (Arrays.isNullOrEmpty(args) || args.length == 1) {
+      return UNDEFINED_VALUE;
+    } else if (args.length == 2) {
+      return args[1];
+    } else {
+      var boundaries = Lists.newArrayList(args);
+      boundaries.remove(0);
+      return translationService.formatList(boundaries);
+    }
   }
 }