+++ mod-record-specifications-server/src/main/java/org/folio/rspec/service/SpecificationFieldService.java
@@ -1,12 +1,17 @@
 package org.folio.rspec.service;
 
 import java.util.UUID;
+import java.util.function.Function;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.folio.rspec.domain.dto.FieldIndicatorChangeDto;
+import org.folio.rspec.domain.dto.FieldIndicatorDto;
+import org.folio.rspec.domain.dto.FieldIndicatorDtoCollection;
 import org.folio.rspec.domain.dto.Scope;
 import org.folio.rspec.domain.dto.SpecificationFieldChangeDto;
 import org.folio.rspec.domain.dto.SpecificationFieldDto;
 import org.folio.rspec.domain.dto.SpecificationFieldDtoCollection;
+import org.folio.rspec.domain.entity.Field;
 import org.folio.rspec.domain.entity.Specification;
 import org.folio.rspec.domain.repository.FieldRepository;
 import org.folio.rspec.exception.ResourceNotFoundException;
@@ -21,6 +26,7 @@ public class SpecificationFieldService {
 
   private final FieldRepository fieldRepository;
   private final SpecificationFieldMapper specificationFieldMapper;
+  private final FieldIndicatorService indicatorService;
 
   public SpecificationFieldDtoCollection findSpecificationFields(UUID specificationId) {
     log.debug("findSpecificationFields::specificationId={}", specificationId);
@@ -54,5 +60,27 @@ public class SpecificationFieldService {
     var updatedField = specificationFieldMapper.partialUpdate(changeDto, fieldEntity);
 
     return specificationFieldMapper.toDto(fieldRepository.save(updatedField));
+  }
+
+  @Transactional
+  public FieldIndicatorDtoCollection findFieldIndicators(UUID fieldId) {
+    log.debug("findFieldIndicators::fieldId={}", fieldId);
+    return doForFieldOrFail(fieldId,
+      field -> indicatorService.findFieldIndicators(fieldId)
+    );
+  }
+
+  @Transactional
+  public FieldIndicatorDto createLocalIndicator(UUID fieldId, FieldIndicatorChangeDto createDto) {
+    log.debug("createLocalIndicator::fieldId={}, createDto={}", fieldId, createDto);
+    return doForFieldOrFail(fieldId,
+      field -> indicatorService.createLocalIndicator(field, createDto)
+    );
+  }
+
+  private <T> T doForFieldOrFail(UUID fieldId, Function<Field, T> action) {
+    return fieldRepository.findById(fieldId)
+      .map(action)
+      .orElseThrow(() -> ResourceNotFoundException.forField(fieldId));
   }
 }