+++ mod-record-specifications-server/src/test/java/org/folio/rspec/controller/SpecificationStorageFieldsControllerTest.java
@@ -1,25 +1,32 @@
 package org.folio.rspec.controller;
 
 import static java.util.UUID.randomUUID;
+import static org.folio.support.ApiEndpoints.fieldIndicatorsPath;
 import static org.folio.support.ApiEndpoints.fieldPath;
 import static org.hamcrest.CoreMatchers.hasItem;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.util.UUID;
 import org.folio.rspec.config.TranslationConfig;
+import org.folio.rspec.domain.dto.FieldIndicatorDto;
+import org.folio.rspec.domain.dto.FieldIndicatorDtoCollection;
 import org.folio.rspec.domain.dto.SpecificationFieldChangeDto;
 import org.folio.rspec.domain.dto.SpecificationFieldDto;
 import org.folio.rspec.exception.ResourceNotFoundException;
 import org.folio.rspec.service.SpecificationFieldService;
+import org.folio.spring.testing.extension.Random;
 import org.folio.spring.testing.extension.impl.RandomParametersExtension;
 import org.folio.spring.testing.type.UnitTest;
 import org.junit.jupiter.api.Test;
@@ -142,6 +149,105 @@ class SpecificationStorageFieldsControllerTest {
       .andExpect(status().isNotFound())
       .andExpect(jsonPath("$.errors.[*].message", hasItem(is("specification with ID [%s] was not found."
         .formatted(specificationId)))));
+  }
+
+  @Test
+  void getIndicators(@Random FieldIndicatorDto indicatorDto) throws Exception {
+    var fieldId = UUID.randomUUID();
+    when(specificationFieldService.findFieldIndicators(fieldId))
+      .thenReturn(new FieldIndicatorDtoCollection().totalRecords(1).addIndicatorsItem(indicatorDto));
+
+    var requestBuilder = get(fieldIndicatorsPath(fieldId))
+      .contentType(APPLICATION_JSON);
+
+    mockMvc.perform(requestBuilder)
+      .andExpect(status().isOk())
+      .andExpect(jsonPath("$.totalRecords", is(1)))
+      .andExpect(jsonPath("$.indicators.size()", is(1)))
+      .andExpect(jsonPath("$.indicators[0].id", is(indicatorDto.getId().toString())))
+      .andExpect(jsonPath("$.indicators[0].fieldId", is(indicatorDto.getFieldId().toString())))
+      .andExpect(jsonPath("$.indicators[0].order", is(indicatorDto.getOrder())))
+      .andExpect(jsonPath("$.indicators[0].label", is(indicatorDto.getLabel())));
+
+    verify(specificationFieldService).findFieldIndicators(fieldId);
+  }
+
+  @Test
+  void createFieldLocalIndicator_createNewLocalIndicator(@Random FieldIndicatorDto indicatorDto) throws Exception {
+    var fieldId = UUID.randomUUID();
+    when(specificationFieldService.createLocalIndicator(eq(fieldId), any())).thenReturn(indicatorDto);
+
+    var requestBuilder = post(fieldIndicatorsPath(fieldId))
+      .contentType(APPLICATION_JSON)
+      .content("{\"order\": 1, \"label\": \"Ind 1\"}");
+
+    mockMvc.perform(requestBuilder)
+      .andExpect(status().isCreated())
+      .andExpect(jsonPath("$.id", is(indicatorDto.getId().toString())))
+      .andExpect(jsonPath("$.fieldId", is(indicatorDto.getFieldId().toString())))
+      .andExpect(jsonPath("$.label", is(indicatorDto.getLabel())))
+      .andExpect(jsonPath("$.order", is(indicatorDto.getOrder())));
+  }
+
+  @Test
+  void createFieldLocalIndicator_return404_notExistingField() throws Exception {
+    var fieldId = UUID.randomUUID();
+    when(specificationFieldService.createLocalIndicator(eq(fieldId), any()))
+      .thenThrow(ResourceNotFoundException.forField(fieldId));
+
+    var requestBuilder = post(fieldIndicatorsPath(fieldId))
+      .contentType(APPLICATION_JSON)
+      .content("{\"order\": 1, \"label\": \"Ind 1\"}");
+
+    mockMvc.perform(requestBuilder)
+      .andExpect(status().isNotFound())
+      .andExpect(jsonPath("$.errors.[*].message", hasItem(is("field definition with ID [%s] was not found."
+        .formatted(fieldId)))));
+  }
+
+  @CsvSource(delimiter = '|', value = {
+    "{\"label\": \"Ind 1m\"}  | order",
+    "{\"order\": 1}           | label"
+  })
+  @ParameterizedTest
+  void createFieldLocalIndicator_return400_missingFieldInPayload(String content, String field) throws Exception {
+    var requestBuilder = post(fieldIndicatorsPath(UUID.randomUUID()))
+      .contentType(APPLICATION_JSON)
+      .content(content);
+
+    mockMvc.perform(requestBuilder)
+      .andExpect(status().isBadRequest())
+      .andExpect(jsonPath("$.errors.[*].message", hasItem(is("Field [%s] must be not null.".formatted(field)))));
+  }
+
+  @Test
+  void createFieldLocalIndicator_return400_highOrder() throws Exception {
+    var requestBuilder = post(fieldIndicatorsPath(UUID.randomUUID()))
+      .contentType(APPLICATION_JSON)
+      .content("{\"order\": 3, \"label\": \"Ind 3\"}");
+
+    mockMvc.perform(requestBuilder)
+      .andExpect(status().isBadRequest())
+      .andExpect(jsonPath("$.errors.[*].message",
+        hasItem(is("order must be less than or equal to 2."))))
+      .andExpect(jsonPath("$.errors.[*].code", hasItem(is("103"))))
+      .andExpect(jsonPath("$.errors.[*].parameters.[*].key", hasItem(is("order"))))
+      .andExpect(jsonPath("$.errors.[*].parameters.[*].value", hasItem(is("3"))));
+  }
+
+  @Test
+  void createFieldLocalIndicator_return400_lowOrder() throws Exception {
+    var requestBuilder = post(fieldIndicatorsPath(UUID.randomUUID()))
+      .contentType(APPLICATION_JSON)
+      .content("{\"order\": 0, \"label\": \"Ind 0\"}");
+
+    mockMvc.perform(requestBuilder)
+      .andExpect(status().isBadRequest())
+      .andExpect(jsonPath("$.errors.[*].message",
+        hasItem(is("order must be greater than or equal to 1."))))
+      .andExpect(jsonPath("$.errors.[*].code", hasItem(is("103"))))
+      .andExpect(jsonPath("$.errors.[*].parameters.[*].key", hasItem(is("order"))))
+      .andExpect(jsonPath("$.errors.[*].parameters.[*].value", hasItem(is("0"))));
   }
 
 }